chore: update to pg release, remove deals
