+++ package.json
@@ -6,7 +6,6 @@
   "types": "dist/index",
   "scripts": {
     "prepublishOnly": "npm run build",
-    "prepare": "npm run compile",
     "prebuild": "npm run clean",
     "build": "npm run compile",
     "compile": "tsc -b tsconfig.json",
@@ -66,6 +65,6 @@
   },
   "dependencies": {
     "@improbable-eng/grpc-web-node-http-transport": "^0.12.0",
-    "@textile/grpc-powergate-client": "0.1.1-pre.2"
+    "@textile/grpc-powergate-client": "0.1.1"
   }
 }
