--- src/deals/index.spec.ts
@@ -1,9 +1,0 @@
-describe("deals", () => {
-  // const c = createDeals({ host, transport: getTransport() })
-  // it("should import", async () => {
-  //   const buffer = fs.readFileSync(`sample-data/samplefile`)
-  //   const { dataCid, size } = await c.import(buffer, false)
-  //   expect(dataCid).length.greaterThan(0)
-  //   expect(size).greaterThan(0)
-  // })
-})
