+++ src/options.ts
@@ -1,4 +1,3 @@
-import * as dealsOptions from "./deals/options"
 import * as ffsOptions from "./ffs/options"
 
-export { ffsOptions, dealsOptions }
+export { ffsOptions }
